+++ src/components/VvAvatar/index.ts
@@ -1,0 +1,9 @@
+import { ModifiersProps } from '@/props'
+
+export const VvAvatarProps = {
+	...ModifiersProps,
+	/**
+	 * Image src for avatar
+	 */
+	imgSrc: String,
+}