+++ src/components/VvButtonGroup/index.ts
@@ -7,10 +7,22 @@ export const VvButtonGroupProps = {
 	/**
 	 * String or String[] of css classes (modifiers) that will be provided to each button'
 	 */
-	itemModifiers: { type: [String, Array<string>], default: undefined },
+	itemModifiers: {
+		type: [String, Array] as PropType<string | string[]>,
+		default: '',
+	},
 	toggle: { type: Boolean, default: false },
 	multiple: { type: Boolean, default: false },
-	modelValue: { type: [String, Array<string>], default: undefined },
+	modelValue: {
+		type: [String, Number, Boolean, Array] as PropType<
+			| string
+			| number
+			| boolean
+			| (string | number | boolean)[]
+			| undefined
+		>,
+		default: undefined,
+	},
 }
 
 export const VvButtonGroupEvents = ['update:modelValue']