+++ src/components/VvCheckboxGroup/VvCheckboxGroup.vue
@@ -56,7 +56,7 @@
 			value: getOptionValue(option),
 		}
 	}
-	const { HintSlot } = HintSlotFactory(props, slots)
+	const { HintSlot, hintSlotScope } = HintSlotFactory(props, slots)
 </script>
 
 <template>
@@ -74,6 +74,19 @@
 			<!-- #endregion -->
 			<slot v-else />
 		</div>
-		<HintSlot class="vv-checkbox-group__hint" />
+		<HintSlot class="vv-checkbox-group__hint">
+			<template v-if="$slots.hint" #hint>
+				<slot name="hint" v-bind="hintSlotScope" />
+			</template>
+			<template v-if="$slots.loading" #loading>
+				<slot name="loading" v-bind="hintSlotScope" />
+			</template>
+			<template v-if="$slots.valid" #valid>
+				<slot name="valid" v-bind="hintSlotScope" />
+			</template>
+			<template v-if="$slots.invalid" #invalid>
+				<slot name="invalid" v-bind="hintSlotScope" />
+			</template>
+		</HintSlot>
 	</fieldset>
 </template>