+++ src/components/VvNavSeparator/VvNavSeparator.ts
@@ -1,0 +1,8 @@
+export default defineComponent({
+	render() {
+		return h('li', {
+			class: 'vv-nav__divider',
+			role: 'separator',
+		})
+	},
+})